packages_py/vault_file/pyproject.toml
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "vault-file"
version = "1.0.0"
description = "Vault File logic for Python"
authors = ["Admin <admin@example.com>"]
packages = [{include = "vault_file", from = "src"}]

[tool.poetry.dependencies]
python = "^3.11"
python-dotenv = ">=1.0.0"
pydantic = ">=2.0.0"
pyyaml = ">=6.0.0"

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
